requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.6
django-allauth==0.63.3
django-widget-tweaks==1.5.0
idna==3.7
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.8.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
